% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hasse.R
\name{hasseGraph}
\alias{hasseGraph}
\title{Create the Hasse graph}
\usage{
hasseGraph(tmsobj, random = NULL, restricted = TRUE, data = NULL)
}
\arguments{
\item{tmsobj}{A terms object.}

\item{random}{A character vector of terms considered random.}

\item{restricted}{should the restricted or unrestricted model be calculated?}

\item{data}{A data frame.}
}
\value{
A Hasse graph.
}
\description{
Create the Hasse graph
}
