% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hasse.R
\name{mixed}
\alias{mixed}
\title{Create the anova table for a mixed model anova}
\usage{
mixed(aov, random = NULL, restricted = TRUE)
}
\arguments{
\item{aov}{and aov object.}

\item{random}{A character vector of terms considered random.}

\item{restricted}{should the restricted or unrestricted model be calculated?}
}
\value{
Aov object with recalculated anova table.
}
\description{
Create the anova table for a mixed model anova
}
