% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hasse.R
\name{hasseDiagram}
\alias{hasseDiagram}
\title{Create a Hasse Diagram.}
\usage{
hasseDiagram(formula, random = NULL, data = NULL)
}
\arguments{
\item{formula}{A formula object.}

\item{random}{A character vector of terms considered random.}

\item{data}{A data frame.}
}
\value{
DiagrammeR script for the Hasse diagram.
}
\description{
Create a Hasse Diagram.
}
